requirements.txt
python-telegram-bot==21.0
httpx==0.27.0
qrcode==7.4.2
wikipedia==1.4.0